Commented out Notifcations Panel, since its untestable right now Removed StatsWindow Added code that was missed during the shift from StatsWindow to Notification Panel
